django_projects/groups_api/requirements.txt
Django
psycopg2-binary
python-decouple
Pillow
djangorestframework
